requirements.txt
fastapi
uvicorn[standard]
python-dotenv
celery
redis
pydantic
gunicorn
google-generativeai